requirements.txt
# Core Agent Framework
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9

# RAG & Vector Store
chromadb==0.5.5
sentence-transformers==3.0.1
PyMuPDF==1.24.9

# Backend
fastapi==0.112.1
uvicorn==0.30.6
python-multipart==0.0.9

# Database (vehicle profiles - PostgreSQL + SQLite fallback)
sqlalchemy==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Utilities
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
